telegram-ai-bot/requirements.txt

python-telegram-bot==20.8
python-dotenv
requests
pint

